Added in function "get_interface_coefficients" variable "n_value", which represents the value of the normal to the intersection (1 or -1); it is then used to evaluate the coefficients for the gradients along x and along y.
Changed last argument pass to  function "get_interface_distances" from
"n_axis" to "n_normal_inter"; from the last one we can evaluate "n_axis"
and also "n_value", used to evaluate positive or negative distances on the
boundary. Also, changed absolute value evaluating distances for the centers
of the owners of the intersection, fixing the vector as distance from the
center with the inner normal to the center with the outer normal. That's
why, outside in the function "get_interface_coefficients", we multiply the
coefficients obtained for the gradients by the new variable "n_value"